Move analysis code into its own module.
